Convert Renderer to React
